This should make versioning part of the commits
The Arduino library setup requires the library.properties file to
contain the current version. So we can't auto-generate that file, it
has to be checked into the repo.

make version VERSION=<version> will update library.properties, tag the
repo and push it which in turn will kick the release process on travis.